backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.9.0
openai>=1.0.0
python-dotenv>=1.0.1
